+++ app/Database/Migrations/20220501100512_Hoomrooms.php
@@ -31,13 +31,13 @@ class Hoomrooms extends Migration
                 'constraint'     => 1,
             ],
             'createdAt' => [
-                'type' => 'int',
+                 'type' => 'bigint',
             ],
             'updatedAt' => [
-                'type' => 'int',
+                 'type' => 'bigint',
             ],
             'deletedAt' => [
-                'type' => 'int', 
+                 'type' => 'bigint', 
             ],
         ]);
         $this->forge->addKey('id', true);
